Added all the functionalities in the app and creates languages files as well

+++ lib/models/event.dart
@@ -17,8 +17,8 @@ class Event {
   String? startTime;
   String? subtitle;
   String? title;
-  int? longitude;
-  int? latitude;
+  double? longitude;
+  double? latitude;
   bool? isfavorite;
 
   Event(
